#fix/reducing bundle size (#43)
* update : installed @next/build-analyzer + reducing _app.tsx bundle, more to come

* update : code splitting page on /index, /article/[slug], /user/posts

* fix : title not showing when lazy loading the whole app

* update : added splash screen

* update : added splash screen

* fix : splash screen getting cut out

* fixing : SplashScreen ui

* update : add analyze to gitignore

* fix : error importing SplashScreen

* update : optimizing bundle + some ui fixes

* update : added sidebar content when user is no logged in yet

* fix : redirect button destination after login

* update : added mobile header more to come.

* update : added placeholder for comment form if user is not logged in

* update : added simple alert when user unauthorized user tyring to react to a comment.

* update : optimizing article progress bar behavior

* update : added transparency to mobile header

* update : more fix on article progress bar behavior

* update : z-index clashing between reading progress bar, mobile header and sidebar

* update : tackled some spacing problem on mobile header

* update : ui update on unauthorized comment panel

* udpate : upgraded daisyui

* update :  fixed z-index problem on sort options

* update : added cross-env package to run next build analyzer on windows

* update : added animated-tailwindcss package

* update : redesign the splash screen

* update :  more ui improvement with transition
