+++ layouts/article/comment/LayoutArticleCommentSection.tsx
@@ -1,14 +1,20 @@
 import { useEffect, useState } from "react";
 import { MdSort } from "react-icons/md";
 import ArticleComments from "../../../components/article/ArticleComments";
+import { useAuthCtx } from "../../../utils/contexts/auth/AuthHook";
 import {
   CommentModelsSortType,
   CommentModelListPagedSorted,
 } from "../../../utils/data/models/CommentModel";
 import { fbCommentGet } from "../../../utils/services/network/FirebaseApi/FirebaseCommentModules";
 import LayoutArticleCommentForm from "./LayoutArticleCommentForm";
+import Link from "next/link";
 
 function LayoutArticleCommentSection({ articleId }: { articleId: string }) {
+  const {
+    authStt: { user },
+  } = useAuthCtx();
+
   const [commentList, setCommentList] = useState<CommentModelListPagedSorted>();
   const [sortedBy, setSortedBy] = useState<CommentModelsSortType>("new");
 
@@ -56,31 +62,32 @@ function LayoutArticleCommentSection({ articleId }: { articleId: string }) {
     loadComments();
   }, [sortedBy]);
 
+  console.log(user);
   // console.log("commentList", commentList);
 
   // console.log("render ArticleCommentSection");
   return (
     <div className="flex flex-col gap-4 sm:gap-8">
       {commentList?.comments.length && (
-        <div className="flex gap-2 sm:gap-4 items-center justify-between sm:justify-start">
-          <span className="font-bold text-sm sm:text-base truncate">{`${commentList.comments.length} Comments`}</span>
-          <div className="dropdown dropdown-end">
+        <div className="flex animate-fadeIn items-center justify-between gap-2 sm:justify-start sm:gap-4 z-[1]">
+          <span className="truncate text-sm font-bold sm:text-base">{`${commentList.comments.length} Comments`}</span>
+          <div className="dropdown-end dropdown">
             <label
               tabIndex={0}
-              className="btn btn-ghost rounded-lg gap-2 sm:gap-4 flex-nowrap "
+              className="btn-ghost btn flex-nowrap gap-2 rounded-xl sm:gap-4 "
             >
               <label className="text-xl sm:text-2xl">
                 <MdSort />
               </label>
-              <span className="text-sm sm:text-base truncate">
+              <span className="truncate text-sm sm:text-base">
                 Sorted by{" "}
                 <span className="underline">{getSortedByLabel()}</span>
               </span>
             </label>
             <ul
               tabIndex={0}
-              className="dropdown-content menu p-2 ring-[1px] ring-base-content bg-base-100 rounded-lg w-52
-            [&_a]:!rounded-lg text-sm sm:text-base font-bold"
+              className="dropdown-content menu !z-[1] w-52 rounded-xl bg-base-300 p-2 text-sm
+              font-bold ring-[1px] ring-base-content/10 sm:text-base [&_a]:!rounded-xl"
             >
               {sorts.map((e, idx) => {
                 return (
@@ -101,13 +108,47 @@ function LayoutArticleCommentSection({ articleId }: { articleId: string }) {
         </div>
       )}
 
-      <LayoutArticleCommentForm
-        articleId={articleId}
-        loadComments={loadComments}
-      />
-      <div></div>
+      {user && (
+        <LayoutArticleCommentForm
+          articleId={articleId}
+          loadComments={loadComments}
+        />
+      )}
+      {!user && (
+        <div
+          className="my-4 flex w-full animate-fadeIn flex-col rounded-xl bg-primary/10 p-4 text-center sm:my-8 sm:p-8"
+        >
+          <div className="flex w-full flex-wrap items-center justify-center gap-4 sm:gap-8">
+            <span className="flex-2 w-full text-lg font-bold  sm:text-xl">
+              Join us now, let us know what do you think about this amazing
+              article!
+            </span>
+            <Link
+              href={{
+                pathname: "/auth",
+              }}
+              passHref
+            >
+              <a className="--btn-resp flex-2 btn-primary btn w-full gap-2 font-bold normal-case sm:flex-[2_2_0%]">
+                Login, if have account
+              </a>
+            </Link>
+            <Link
+              href={{
+                pathname: "/auth",
+              }}
+              passHref
+            >
+              <a className="--btn-resp flex-2 btn-primary btn w-full gap-2 font-bold normal-case sm:flex-[2_2_0%]">
+                {`Register, if you don't`}
+              </a>
+            </Link>
+          </div>
+        </div>
+      )}
+
       {!commentList?.comments.length ? (
-        <span className="font-bold text-sm sm:text-base  text-center opacity-80 my-16">
+        <span className="my-16 text-center text-sm  font-bold opacity-80 sm:text-base animate-fadeIn">
           {`No comments yet, let the people know, what's your thought about this article..`}
         </span>
       ) : (
