+++ layouts/auth/pages/LayoutAuthPage.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import AuthPanel from "../../../components/auth/AuthPanel";
+import MainContainer from "../../../components/main/MainContainer";
+
+function LayoutAuthPage() {
+  return (
+    <MainContainer>
+      <AuthPanel />
+    </MainContainer>
+  );
+}
+
+export default LayoutAuthPage;
