+++ next.config.js
@@ -5,3 +5,9 @@ const nextConfig = {
 }
 
 module.exports = nextConfig
+
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+  enabled: process.env.ANALYZE === "true",
+});
+
+module.exports = withBundleAnalyzer({});
