+++ package.json
@@ -6,14 +6,16 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "scan": "cross-env ANALYZE=true npm run build"
   },
   "dependencies": {
     "@headlessui/react": "^1.6.4",
+    "@next/bundle-analyzer": "^12.3.0",
     "@tailwindcss/line-clamp": "^0.4.0",
     "@types/ua-parser-js": "^0.7.36",
     "axios": "^0.27.2",
-    "daisyui": "^2.13.6",
+    "daisyui": "^2.27.0",
     "date-fns": "^2.29.1",
     "dotenv": "^16.0.1",
     "firebase": "^9.8.3",
@@ -38,6 +40,7 @@
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.4",
     "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "animated-tailwindcss": "^3.3.0",
     "autoprefixer": "^10.4.7",
     "eslint": "8.15.0",
     "eslint-config-next": "^12.2.5",
