+++ pages/profile/choosetopics.tsx
@@ -1,8 +1,17 @@
+import dynamic from "next/dynamic";
 import Head from "next/head";
-import React from "react";
-import MainContainer from "../../components/main/MainContainer";
-import ProfileChooseTopicsForm from "../../components/profile/ProfileChooseTopicsForm";
+import SplashScreen from "../../components/placeholder/SplashScreen";
 import { APP_NAME } from "../../utils/helpers/Constants";
+
+const LazyLayoutProfilePageChooseTopics = dynamic(
+  () => import("../../layouts/profile/pages/LayoutProfilePageChooseTopics"),
+  {
+    ssr: false,
+    loading(loadingProps) {
+      return <SplashScreen />;
+    },
+  },
+);
 
 function ProfileChooseTopics() {
   const title = "Setup your " + APP_NAME + " profile - " + APP_NAME;
@@ -14,13 +23,8 @@ function ProfileChooseTopics() {
         <meta name="viewport" content="initial-scale=1.0, width=device-width" />
         <meta name="description" content={"Setup your profile"} />
       </Head>
-      {/* <Header /> */}
-      <MainContainer>
-        <div className="relative z-0 min-h-screen overflow-hidden rounded-xl">
-        <ProfileChooseTopicsForm />
-        </div>
-      </MainContainer>
-      {/* <Footer/> */}
+      
+      <LazyLayoutProfilePageChooseTopics />
     </>
   );
 }
