+++ pages/write.tsx
@@ -1,9 +1,17 @@
+import dynamic from "next/dynamic";
 import Head from "next/head";
-import React from "react";
-import MainContainer from "../components/main/MainContainer";
-import WritingPanel from "../components/write/WritingPanel";
-import { WritingPanelProvider } from "../utils/contexts/writingPanel/WritingPanelProvider";
+import SplashScreen from "../components/placeholder/SplashScreen";
 import { APP_DESC, APP_NAME } from "../utils/helpers/Constants";
+
+const LazyLayoutArticlePageWrite = dynamic(
+  () => import("../layouts/article/pages/LayoutArticlePageWrite"),
+  {
+    ssr: false,
+    loading(loadingProps) {
+      return <SplashScreen />;
+    },
+  },
+);
 
 function Write() {
   return (
@@ -13,11 +21,7 @@ function Write() {
         <meta name="viewport" content="initial-scale=1.0, width=device-width" />
         <meta name="description" content={APP_DESC} />
       </Head>
-      <MainContainer className="">
-        <WritingPanelProvider>
-          <WritingPanel />
-        </WritingPanelProvider>
-      </MainContainer>
+      <LazyLayoutArticlePageWrite />
     </>
   );
 }
